add docker  and tests

+++ back-end-dot-net/DsDeliveryApi/DsDeliveryApi/Repositories/IOrderRepository.cs
@@ -12,7 +12,7 @@ namespace DsDeliveryApi.Repositories
         Task<List<Order>> FindAsync(Expression<Func<Order, bool>> predicate);
         Task<List<Order>> FindOrdersWithProducts();
         Task<List<Order>> GetAllAsync();
-        Task<Order> GetByIdAsync(long id);
+        Task<Order> GetByIdAsync(int id);
         Task<bool> RemoveAsync(Order entity);
         Task<Order> UpdateAsync(Order entity);
     }
